move files that do not match their classname to make zeitwerk happy
