+++ config/application.rb
@@ -205,10 +205,7 @@ if ["test", "development"].include?(Rails.env)
 end
 
 require 'samson/hooks'
-
-require_relative "../lib/samson/syslog_formatters"
-require_relative "../lib/samson/logging"
-require_relative "../lib/samson/initializer_logging"
+require_relative "logging"
 require_relative "../app/models/job_queue" # need to load early or dev reload will lose the .enabled
 
 # prevents `Unknown validator: 'Doorkeeper::RedirectUriValidator'`
