+++ lib/samson/syslog_formatter.rb
