+++ lib/warden/strategies/doorkeeper.rb
