--- test/lib/samson/initializer_logging_test.rb
@@ -1,7 +1,0 @@
-# frozen_string_literal: true
-require_relative '../../test_helper'
-
-SingleCov.covered! uncovered: 4
-
-describe 'InitializerLogging' do
-end
