--- test/lib/samson/logging_test.rb
@@ -1,7 +1,0 @@
-# frozen_string_literal: true
-require_relative '../../test_helper'
-
-SingleCov.not_covered!
-
-describe 'Logging' do
-end
