+++ test/lib/samson/syslog_formatter_test.rb
