+++ test/lib/warden/strategies/doorkeeper_test.rb
